Added easing event and fixed currentLag
Added an easing event (was running tests and was useful to have), and a new configuration method, consecutiveLags(int), which allows one to configure the number of lag spikes required before the onLag() function is triggered.
Additionally, a change to line #186 was made, as the currentLag was being set to NaN as a result of currentlag being set to a number multiplied by 0 (initial currentLag) in previous iterations. Calculations can now occur properly.